+++ xbmc/cores/VideoPlayer/DVDInputStreams/InputStreamPVRBase.cpp
@@ -293,7 +293,7 @@ void CInputStreamPVRBase::UpdateStreamMap()
       streamVideo->iFpsRate = stream.iFPSRate;
       streamVideo->iHeight = stream.iHeight;
       streamVideo->iWidth = stream.iWidth;
-      streamVideo->fAspect = stream.fAspect;
+      streamVideo->fAspect = static_cast<double>(stream.fAspect);
 
       dStream = streamVideo;
     }
