+++ xbmc/windowing/osx/VideoSyncOsx.mm
@@ -100,7 +100,7 @@ void CVideoSyncOsx::VblankHandler(int64_t nowtime, uint32_t timebase)
   if (m_LastVBlankTime != 0)
   {
     VBlankTime = (double)(nowtime - m_LastVBlankTime) / (double)timebase;
-    NrVBlanks = MathUtils::round_int(VBlankTime * m_fps);
+    NrVBlanks = MathUtils::round_int(VBlankTime * static_cast<double>(m_fps));
 
     //update the vblank timestamp, update the clock and send a signal that we got a vblank
     UpdateClock(NrVBlanks, Now, m_refClock);
